weed
bud-core (B.U.D. 1.0/2.0/3.0, 103 modul) CI'da hic derlenmiyordu: ci.yml'e uc yapilandirma (default/bud3/no-default) clippy -D warnings + test adimi baglandi; clippy 91 ihlal -> 0; exepdf.rs if'in iki dali ayni idi (icerik-duyarli bolumleme iddiasi sahte, olcum hesaplanip atiliyordu) sabit oranli hale getirildi; logfield.rs vec![Vec::with_capacity(n); 7] kapasiteyi 7 sutunun 6'sinda kaybediyordu; lib.rs cift cfg yuzunden bud_format_qrvideo ve bud_format_sertlesme 3.0 kapisinin disinda kalmisti (bud2 derlemesine 3.0 kodu siziyordu); pgm.rs never_loop tek gecisli dal if'e cevrildi; securededup.rs explicit_counter_loop; revolutionary.rs inherent to_string -> Display; price.rs media_device_only_holds assert!(0.0 <= 0.016) iki sabiti karsilastiriyordu, model uzerinden gercek teste cevrildi + kanarya; test_bud2.rs assert!(true) yer tutucusu 6 kayipsizlik degismezi testiyle degistirildi (mutasyon kanaryasi ile dogrulandi: tek bit bozulmada 4 test kirmizi); bud/Cargo.toml [lints.rust]/[lints.clippy] kapisi + bud1/bud2 feature'larinin cfg kapisi OLMADIGI belgelendi; 490/490 test gecer

+++ bud/src/bud_format_timeseries.rs
@@ -2,8 +2,8 @@
 //!
 //! K92: zaman serisi sıkıştırması 10-12x (ardışık zaman damgası farkı
 //! + XOR kayan nokta değerleri). B.U.D. telemetri/ölçüm verisi için domain transformu:
-//! (ts, value) çiftlerini zaman farkı + XOR bit akışına çevirir - zstd'nin
-//! göremediği yüksek entropili float farklarını görür.
+//!   (ts, value) çiftlerini zaman farkı + XOR bit akışına çevirir - zstd'nin
+//!   göremediği yüksek entropili float farklarını görür.
 //!
 //! Kayıpsız: encode → decode = orijinal (K38). Panik'siz, no unsafe, deterministik.
 //!
@@ -107,12 +107,12 @@ impl TimeSeriesColumnar {
             } else if (-63..=63).contains(&delta) {
                 w.write_bit(true);
                 w.write_bit(false);
-                w.write_bits((delta as i64 + 63) as u64, 7);
+                w.write_bits((delta + 63) as u64, 7);
             } else if (-255..=255).contains(&delta) {
                 w.write_bit(true);
                 w.write_bit(true);
                 w.write_bit(false);
-                w.write_bits((delta as i64 + 255) as u64, 9);
+                w.write_bits((delta + 255) as u64, 9);
             } else {
                 w.write_bit(true);
                 w.write_bit(true);
